refactor: seperate the flight form and modal
